Move CheckpointStatus to stream sub package and make it package protected
